fix: align MaxWeightKg C# type with SQL DECIMAL(18,2)
- Change StorageCell.MaxWeightKg from double to decimal
- Update ViewModel, DatabaseService, DbContext and migrations
- Fixes 'Unable to cast Decimal to Double' on fresh DB install

+++ AISPVZ/Models/Employee.cs
@@ -69,7 +69,7 @@ public class StorageCell
 
     public bool IsBusy { get; set; } = false;
 
-    public double MaxWeightKg { get; set; } = 30.0;
+    public decimal MaxWeightKg { get; set; } = 30.0m;
 
     [MaxLength(200)]
     public string? Comment { get; set; }
